end of tutorial

+++ app/Http/Controllers/API/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UsersController extends Controller
@@ -26,7 +27,9 @@ class UsersController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        if (Gate::allows('isAdmin') || Gate::allows('isAuthor')) {
+            return User::latest()->paginate(5);
+        }
     }
 
     /**
@@ -51,7 +54,56 @@ class UsersController extends Controller
             'bio' => $request->bio
         ]);
     }
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function profile()
+    {
+        return auth('api')->user();
+    }
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+        $this->validate($request, [
+            'name' => 'required|string|max:191',
+            'email' => 'required|email|max:191|unique:users,email,' . $user->id,
+            'password' => 'sometimes|min:6',
+        ]);
+        $currentPhoto = $user->photo; //defalut photo.png in database
 
+        // if user upladed a photo
+        if ($request->photo !== $currentPhoto) {
+            // take the string from photo name and replace it witn a unique number
+            $name = time() . '.' .
+                explode('/', explode(
+                    ':',
+                    substr($request->photo, 0, strpos($request->photo, ';'))
+                )[1])[1];
+            // store the image in profile folder
+            \Image::make($request->photo)->save(public_path('images/profile/') . $name);
+
+            $request->merge(['photo' => $name]);
+
+            $userPhoto = public_path('images/profile/') . $currentPhoto;
+            if (file_exists($userPhoto)) {
+                @unlink($userPhoto);
+            }
+        };
+
+        if (!empty($request->password)) {
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+        $user->update($request->all());
+    }
     /**
      * Display the specified resource.
      *
@@ -92,8 +144,19 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return ['message' => 'User Deleted'];
+    }
+    public function search()
+    {
+        if ($search = \Request::get('q')) {
+            $users = User::where(function ($query) use ($search) {
+                $query->where('name', 'LIKE', "%$search%")
+                    ->orWhere('email', 'LIKE', "%$search%")->orWhere('type', 'LIKE', "%$search%");
+            })->paginate(20);
+        }
+        return $users;
     }
 }
